Bite-sized refactor: Move all the args in TermControl to their own file. Convert some older macro uses to TYPED_EVENT
