+++ src/cascadia/TerminalControl/EventArgs.cpp
@@ -1,0 +1,12 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+#include "pch.h"
+#include "EventArgs.h"
+#include "TitleChangedEventArgs.g.cpp"
+#include "CopyToClipboardEventArgs.g.cpp"
+#include "PasteFromClipboardEventArgs.g.cpp"
+#include "OpenHyperlinkEventArgs.g.cpp"
+#include "NoticeEventArgs.g.cpp"
+#include "ScrollPositionChangedArgs.g.cpp"
+#include "RendererWarningArgs.g.cpp"
