+++ src/cascadia/TerminalControl/IDirectKeyListener.idl
@@ -1,0 +1,14 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+namespace Microsoft.Terminal.Control
+{
+    // C++/winrt makes it difficult to share this idl between two projects,
+    // Instead, we just pin the uuid and include it in both TermControl and App
+    // If you update this one, please update TerminalApp\IDirectKeyListener.idl.
+    // If you change this interface, please update the guid.
+    // If you press F7 or Alt and get a runtime error, go make sure both copies are the same.
+    [uuid("0ddf4edc-3fda-4dee-97ca-a417ee3dd510")] interface IDirectKeyListener {
+        Boolean OnDirectKeyEvent(UInt32 vkey, UInt8 scanCode, Boolean down);
+    };
+}
