+++ src/cascadia/TerminalControl/TermControl.cpp
@@ -704,7 +704,7 @@ namespace winrt::Microsoft::Terminal::Control::implementation
             }
 
             auto noticeArgs = winrt::make<NoticeEventArgs>(NoticeLevel::Warning, std::move(message));
-            control->_raiseNoticeHandlers(*control, std::move(noticeArgs));
+            control->_RaiseNoticeHandlers(*control, std::move(noticeArgs));
         }
     }
 
@@ -2134,7 +2134,7 @@ namespace winrt::Microsoft::Terminal::Control::implementation
                 co_await strong->Dispatcher();
 
                 auto noticeArgs = winrt::make<NoticeEventArgs>(NoticeLevel::Warning, std::move(msg));
-                strong->_raiseNoticeHandlers(*strong, std::move(noticeArgs));
+                strong->_RaiseNoticeHandlers(*strong, std::move(noticeArgs));
             }();
         }
 
@@ -2399,7 +2399,8 @@ namespace winrt::Microsoft::Terminal::Control::implementation
 
     void TermControl::_TerminalTitleChanged(const std::wstring_view& wstr)
     {
-        _titleChangedHandlers(winrt::hstring{ wstr });
+        auto titleArgs = winrt::make_self<TitleChangedEventArgs>(winrt::hstring{ wstr });
+        _TitleChangedHandlers(*this, *titleArgs);
     }
     void TermControl::_TerminalTabColorChanged(const std::optional<til::color> /*color*/)
     {
@@ -2409,7 +2410,7 @@ namespace winrt::Microsoft::Terminal::Control::implementation
     void TermControl::_CopyToClipboard(const std::wstring_view& wstr)
     {
         auto copyArgs = winrt::make_self<CopyToClipboardEventArgs>(winrt::hstring(wstr));
-        _clipboardCopyHandlers(*this, *copyArgs);
+        _CopyToClipboardHandlers(*this, *copyArgs);
     }
 
     // Method Description:
@@ -2540,7 +2541,7 @@ namespace winrt::Microsoft::Terminal::Control::implementation
                                                                    winrt::to_hstring(htmlData),
                                                                    winrt::to_hstring(rtfData),
                                                                    formats);
-        _clipboardCopyHandlers(*this, *copyArgs);
+        _CopyToClipboardHandlers(*this, *copyArgs);
         return true;
     }
 
@@ -2554,7 +2555,7 @@ namespace winrt::Microsoft::Terminal::Control::implementation
         auto pasteArgs = winrt::make_self<PasteFromClipboardEventArgs>(clipboardDataHandler);
 
         // send paste event up to TermApp
-        _clipboardPasteHandlers(*this, *pasteArgs);
+        _PasteFromClipboardHandlers(*this, *pasteArgs);
     }
 
     // Method Description:
@@ -3226,7 +3227,7 @@ namespace winrt::Microsoft::Terminal::Control::implementation
         co_await Dispatcher();
 
         auto hyperlinkArgs = winrt::make_self<OpenHyperlinkEventArgs>(heldUri);
-        _openHyperlinkHandlers(*strongThis, *hyperlinkArgs);
+        _OpenHyperlinkHandlers(*strongThis, *hyperlinkArgs);
     }
 
     // Method Description:
@@ -3271,7 +3272,7 @@ namespace winrt::Microsoft::Terminal::Control::implementation
     winrt::fire_and_forget TermControl::TaskbarProgressChanged()
     {
         co_await resume_foreground(Dispatcher(), CoreDispatcherPriority::High);
-        _setTaskbarProgressHandlers(*this, nullptr);
+        _SetTaskbarProgressHandlers(*this, nullptr);
     }
 
     // Method Description:
@@ -3317,7 +3318,7 @@ namespace winrt::Microsoft::Terminal::Control::implementation
         if (auto control{ weakThis.get() })
         {
             auto noticeArgs = winrt::make<NoticeEventArgs>(NoticeLevel::Info, RS_(L"TermControlReadOnly"));
-            control->_raiseNoticeHandlers(*control, std::move(noticeArgs));
+            control->_RaiseNoticeHandlers(*control, std::move(noticeArgs));
         }
     }
 
@@ -3392,14 +3393,6 @@ namespace winrt::Microsoft::Terminal::Control::implementation
     // -------------------------------- WinRT Events ---------------------------------
     // Winrt events need a method for adding a callback to the event and removing the callback.
     // These macros will define them both for you.
-    DEFINE_EVENT(TermControl, TitleChanged, _titleChangedHandlers, Control::TitleChangedEventArgs);
     DEFINE_EVENT(TermControl, FontSizeChanged, _fontSizeChangedHandlers, Control::FontSizeChangedEventArgs);
     DEFINE_EVENT(TermControl, ScrollPositionChanged, _scrollPositionChangedHandlers, Control::ScrollPositionChangedEventArgs);
-
-    DEFINE_EVENT_WITH_TYPED_EVENT_HANDLER(TermControl, PasteFromClipboard, _clipboardPasteHandlers, Control::TermControl, Control::PasteFromClipboardEventArgs);
-    DEFINE_EVENT_WITH_TYPED_EVENT_HANDLER(TermControl, CopyToClipboard, _clipboardCopyHandlers, Control::TermControl, Control::CopyToClipboardEventArgs);
-    DEFINE_EVENT_WITH_TYPED_EVENT_HANDLER(TermControl, OpenHyperlink, _openHyperlinkHandlers, Control::TermControl, Control::OpenHyperlinkEventArgs);
-    DEFINE_EVENT_WITH_TYPED_EVENT_HANDLER(TermControl, SetTaskbarProgress, _setTaskbarProgressHandlers, Control::TermControl, IInspectable);
-    DEFINE_EVENT_WITH_TYPED_EVENT_HANDLER(TermControl, RaiseNotice, _raiseNoticeHandlers, Control::TermControl, Control::NoticeEventArgs);
-    // clang-format on
 }
